+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/api/DeliveryServiceApi.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.DeliveryDto;
 import ru.yandex.practicum.dto.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "delivery", path = "/api/v1/delivery")
@@ -17,7 +18,7 @@ public interface DeliveryServiceApi {
     public ResponseEntity<DeliveryDto> createOrUpdateDelivery(@RequestBody DeliveryDto dto);
 
     @PostMapping("/cost")
-    public ResponseEntity<Double> calculateDeliveryCost(@RequestBody OrderDto dto);
+    public ResponseEntity<BigDecimal> calculateDeliveryCost(@RequestBody OrderDto dto);
 
     @PostMapping("/successful")
     public ResponseEntity<DeliveryDto> markDeliverySuccessful(@RequestBody UUID orderId);
